Pridaná trieda DbNode
