+++ src/Data/AbsNode.cpp
@@ -1,8 +1,35 @@
 #include "Data/AbsNode.h"
 #include <QString>
 
-Data::AbsNode::AbsNode( qlonglong id, QString name)
+Data::AbsNode::AbsNode( qlonglong id, QString name, Data::Type* type, Data::Graph* graph)
 {
     this->id=id;
     this->name=name;
+    this->type=type;
+    this->graph = graph;
 }
+
+bool Data::AbsNode::equals( Node* node )
+{
+    //porovnanie s inym uzlom
+    if ( this == node ) {
+        return true;
+    }
+    if ( node == NULL ) {
+        return false;
+    }
+    if ( id != node->getId() ) {
+        return false;
+    }
+    if ( ( graph==NULL && node->getGraph()!=NULL ) || ( graph!=NULL && node->getGraph()==NULL ) ) {
+        return false;
+    }
+    if ( graph==NULL && node->getGraph()==NULL ) {
+        return true;
+    }
+    if ( graph!=NULL && graph->getId() != node->getGraph()->getId() ) {
+        return false;
+    }
+    return true;
+}
+
