+++ src/Data/DbNode.cpp
@@ -1,0 +1,6 @@
+#include "Data/DbNode.h"
+
+Data::DbNode::DbNode()
+{
+    this->inDB = false;
+}
